daemon: extract settings views module from config.rs (#41 cont)
Pulls the get_*/set_* settings-view pairs (LLM, embeddings, persistence,
database, backend tasks, ws auth, connector defaults, api key) into
config/views.rs. Helpers they call are now pub(super); resolution +
default helpers stay in config/mod.rs.

mod.rs is now 2912 lines (from 4373 before #41 began).
